Added The Remaining Player Units And Enemies
Created the remaining four player unit and enemy prefabs respectively | Adjusted variables in the managers to be floats for more flexible stats

+++ Assets/Scripts/Managers/BaseManager.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BaseManager : MonoBehaviour
 {
-    public int baseHealth = 10;
+    public float baseHealth = 10;
 
     public GameObject youWinScreen;
     public GameObject youLoseScreen;
